third_party/WebKit/LayoutTests/fast/dom/node-prefix-setter-namespace-exception-expected.txt

CONSOLE WARNING: Setting 'Element.prefix' is deprecated, as it is read-only per DOM (http://dom.spec.whatwg.org/#element).
Test how Node.prefix setter raises NAMESPACE_ERR.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.createElementNS(null, 'attr').prefix = 'abc' threw exception NamespaceError: Failed to set the 'prefix' property on 'Element': No namespace is set, so a namespace prefix may not be set..
PASS document.createElementNS('foo', 'bar').prefix = 'xml' threw exception NamespaceError: Failed to set the 'prefix' property on 'Element': The prefix 'xml' may not be set on namespace 'foo'..
PASS successfullyParsed is true

TEST COMPLETE
